Moved basic map init into useMap
